backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
pymysql>=1.0.3
alembic>=1.10.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
celery>=5.2.7
redis>=4.5.4
minio>=7.1.14
loguru>=0.7.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
autogen>=0.2.0
langchain>=0.0.267
openai>=0.27.8
tiktoken>=0.4.0
numpy>=1.24.3
pandas>=2.0.1
pytest>=7.3.1
httpx>=0.24.0
tenacity>=8.2.2
PyPDF2>=3.0.1
docx2txt>=0.8
python-docx>=0.8.11
